requirements.txt
numpy==1.23.5
pandas==2.3.1
matplotlib==3.10.5
seaborn==0.13.2
tqdm==4.67.1
PyYAML==6.0.2
mlflow==2.14.2
typer==0.12.3
fastapi==0.115.6
uvicorn[standard]==0.30.6
pytest==8.3.3
httpx==0.27.2
prometheus-client==0.20.0